Fix compiler warnings

+++ src/drivers/src/stm32f407xx_gpio_driver.c
@@ -291,11 +291,11 @@ void GPIO_WriteToOutputPin(GPIO_RegDef_t *pGPIOx, uint8_t PinNumber, uint8_t Val
 	if(Value == GPIO_PIN_SET)
 	{
 		//write 1 to the output data register at the bit field corresponding to the pin number
-		pGPIOx->ODR |= ( 1 << PinNumber);
+		pGPIOx->ODR |= ( 1U << PinNumber);
 	}else
 	{
 		//write 0
-		pGPIOx->ODR &= ~( 1 << PinNumber);
+		pGPIOx->ODR &= ~( 1U << PinNumber);
 	}
 }
 
@@ -363,33 +363,33 @@ void GPIO_IRQInterruptConfig(uint8_t IRQNumber, uint8_t EnorDi)
 		if(IRQNumber <= 31)
 		{
 			//program ISER0 register
-			*NVIC_ISER0 |= ( 1 << IRQNumber );
+			*NVIC_ISER0 |= ( 1U << IRQNumber );
 
 		}else if(IRQNumber > 31 && IRQNumber < 64 ) //32 to 63
 		{
 			//program ISER1 register
-			*NVIC_ISER1 |= ( 1 << (IRQNumber % 32) );
+			*NVIC_ISER1 |= ( 1U << (IRQNumber % 32) );
 		}
 		else if(IRQNumber >= 64 && IRQNumber < 96 )
 		{
 			//program ISER2 register //64 to 95
-			*NVIC_ISER2 |= ( 1 << (IRQNumber % 64) );
+			*NVIC_ISER2 |= ( 1U << (IRQNumber % 64) );
 		}
 	}else
 	{
 		if(IRQNumber <= 31)
 		{
 			//program ICER0 register
-			*NVIC_ICER0 |= ( 1 << IRQNumber );
+			*NVIC_ICER0 |= ( 1U << IRQNumber );
 		}else if(IRQNumber > 31 && IRQNumber < 64 )
 		{
 			//program ICER1 register
-			*NVIC_ICER1 |= ( 1 << (IRQNumber % 32) );
+			*NVIC_ICER1 |= ( 1U << (IRQNumber % 32) );
 		}
 		else if(IRQNumber >= 64 && IRQNumber < 96 )
 		{
 			//program ICER2 register
-			*NVIC_ICER2 |= ( 1 << (IRQNumber % 64) );
+			*NVIC_ICER2 |= ( 1U << (IRQNumber % 64) );
 		}
 	}
 
@@ -415,9 +415,9 @@ void GPIO_IRQPriorityConfig(uint8_t IRQNumber,uint32_t IRQPriority)
 {
 	//1. first lets find out the ipr register
 	uint8_t iprx = IRQNumber / 4;
-	uint8_t iprx_section  = IRQNumber %4 ;
+	uint8_t iprx_section  = IRQNumber % 4 ;
 
-	uint8_t shift_amount = ( 8 * iprx_section) + ( 8 - NO_PR_BITS_IMPLEMENTED) ;
+	uint8_t shift_amount = (uint8_t) ( 8 * iprx_section) + ( 8 - NO_PR_BITS_IMPLEMENTED) ;
 
 	*(  NVIC_PR_BASE_ADDR + iprx ) |=  ( IRQPriority << shift_amount );
